+++ eosq/src/helpers/account.helpers.ts
@@ -216,13 +216,14 @@ export function getAccountResources(account: Account, stakes: StakeDetail[]): Ac
   const refundRequest = account.refund_request
   const rexTokens = account.rex_balance
     ? account.rex_balance.vote_stake
-    : `0.0000 ${Config.price_ticker_name}`
+    : `0.0000 ${Config.chain_core_symbol_code}`
   const rexFunds = account.rex_funds
     ? account.rex_funds.balance
-    : `0.0000 ${Config.price_ticker_name}`
+    : `0.0000 ${Config.chain_core_symbol_code}`
   const rexCpuLoans = account.cpu_loans ? account.cpu_loans : 0
   const rexNetLoans = account.net_loans ? account.net_loans : 0
-  const unit = extractValueWithUnits(totalResources.cpu_weight)[1] || ` ${Config.price_ticker_name}`
+  const unit =
+    extractValueWithUnits(totalResources.cpu_weight)[1] || ` ${Config.chain_core_symbol_code}`
   let stakedCpu = parseFloat(extractValueWithUnits(totalResources.cpu_weight)[0])
   const availableFunds = parseFloat(extractValueWithUnits(account.core_liquid_balance)[0])
   const selfStakedCpu = parseFloat(extractValueWithUnits(selfDelegated.cpu_weight)[0])